requirements.txt
pandas==2.1.4
apache-airflow-providers-postgres==5.7.1
psycopg2-binary==2.9.9
sqlalchemy==1.4.53